Fix potential overflow in MergeWithChannelSelector
When calculating the number of merged MACs per pixel, multiple 32-bit
values are multiplied and the result is put in a 64-bit integer. The
32-bit values will be multiplied by using a 32-bit multiplication that
may overflow before the value is put in the 64-bit integer.

To ensure that the full 64-bit size is used for the result, the first
value is now cast to 64-bit so a 64-bit multiplication is performed
instead.

+++ driver/support_library/src/cascading/McePart.cpp
@@ -952,7 +952,7 @@ bool McePart::MergeWithChannelSelectorBefore(const utils::ConstTensorData& chann
     // If not, then merging may make performance worse!
     uint64_t separateMacsPerPixel =
         GetNumElements(channelSelectorWeights.GetShape()) + GetNumElements(m_WeightsInfo.m_Dimensions);
-    uint64_t mergedMacsPerPixel = m_WeightsInfo.m_Dimensions[0] * m_WeightsInfo.m_Dimensions[1] *
+    uint64_t mergedMacsPerPixel = static_cast<uint64_t>(m_WeightsInfo.m_Dimensions[0]) * m_WeightsInfo.m_Dimensions[1] *
                                   channelSelectorWeights.GetShape()[2] * m_WeightsInfo.m_Dimensions[3];
     if (separateMacsPerPixel < mergedMacsPerPixel)
     {
@@ -1022,7 +1022,7 @@ bool McePart::MergeWithChannelSelectorAfter(const utils::ConstTensorData& channe
     // If not, then merging may make performance worse!
     uint64_t separateMacsPerPixel =
         GetNumElements(m_WeightsInfo.m_Dimensions) + GetNumElements(channelSelectorWeights.GetShape());
-    uint64_t mergedMacsPerPixel = m_WeightsInfo.m_Dimensions[0] * m_WeightsInfo.m_Dimensions[1] *
+    uint64_t mergedMacsPerPixel = static_cast<uint64_t>(m_WeightsInfo.m_Dimensions[0]) * m_WeightsInfo.m_Dimensions[1] *
                                   m_WeightsInfo.m_Dimensions[2] * channelSelectorWeights.GetShape()[3];
     if (separateMacsPerPixel < mergedMacsPerPixel)
     {
